rabbitmq_worker/requirements.txt
pika==1.2.1
Pillow==9.1.1
pytesseract==0.3.9
